# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mithra AI Family Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List user's chats
      parameters:
        - name: folder
          in: query
          schema:
            type: ["string", "null"]
        - name: archived
          in: query
          schema:
            type: ["boolean", "null"]
        - name: pinned
          in: query
          schema:
            type: ["boolean", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a new chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "201":
          description: Created chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /chats/{chatId}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get chat by ID
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "404":
          description: Not found
    patch:
      operationId: updateChat
      tags: [chats]
      summary: Update chat (rename, pin, archive, folder)
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatUpdate"
      responses:
        "200":
          description: Updated chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
    delete:
      operationId: deleteChat
      tags: [chats]
      summary: Delete a chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List user's files
      parameters:
        - name: folder
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserFile"
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileInput"
      responses:
        "201":
          description: Uploaded file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserFile"
  /files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get file by ID
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserFile"
        "404":
          description: Not found
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Chat:
      type: object
      required: [id, userId, title, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        model:
          type: ["string", "null"]
        folder:
          type: ["string", "null"]
        isPinned:
          type: boolean
        isArchived:
          type: boolean
        messageCount:
          type: integer
        lastMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ChatInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        model:
          type: string
        folder:
          type: ["string", "null"]
    ChatUpdate:
      type: object
      properties:
        title:
          type: string
        model:
          type: ["string", "null"]
        folder:
          type: ["string", "null"]
        isPinned:
          type: boolean
        isArchived:
          type: boolean

    # ─── Message ────────────────────────────────────────────────────────────
...
    UserFile:
      type: object
      required: [id, userId, name, type, size, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        type:
          type: string
        mimeType:
          type: ["string", "null"]
        size:
          type: integer
        folder:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        createdAt:
          type: string
    FileInput:
      type: object
      required: [name, type, size]
      properties:
        name:
          type: string
        type:
          type: string
        mimeType:
          type: ["string", "null"]
        size:
          type: integer
        folder:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
...